test: unifying testing across client implementation
