+++ extras/http-client-android/src/test/java/io/a2a/client/http/JdkA2AHttpClientIntegrationTest.java
@@ -1,0 +1,9 @@
+package io.a2a.client.http;
+
+public class JdkA2AHttpClientIntegrationTest extends AbstractA2AHttpClientIntegrationTest {
+
+    @Override
+    protected A2AHttpClient createClient() {
+        return new JdkA2AHttpClient();
+    }
+}
